fix failed test and add more tests
