+++ flink-table/flink-table-planner-blink/src/main/scala/org/apache/flink/table/planner/plan/schema/TableSourceTable.scala
@@ -98,6 +98,6 @@ class TableSourceTable(
       isStreamingMode,
       catalogTable,
       dynamicOptions,
-      newExtraDigests ++ newExtraDigests)
+      extraDigests ++ newExtraDigests)
   }
 }
