+++ flink-table/flink-table-planner-blink/src/test/scala/org/apache/flink/table/planner/runtime/batch/sql/TableSourceITCase.scala
@@ -18,24 +18,15 @@
 
 package org.apache.flink.table.planner.runtime.batch.sql
 
-import org.apache.flink.api.common.typeinfo.TypeInformation
-import org.apache.flink.api.java.typeutils.RowTypeInfo
-import org.apache.flink.table.api.{DataTypes, TableSchema, Types}
-import org.apache.flink.table.planner.runtime.utils.BatchAbstractTestBase.TEMPORARY_FOLDER
+import org.apache.flink.table.planner.factories.TestValuesTableFactory
 import org.apache.flink.table.planner.runtime.utils.BatchTestBase.row
 import org.apache.flink.table.planner.runtime.utils.{BatchTestBase, TestData}
-import org.apache.flink.table.planner.utils.{TestDataTypeTableSource, TestFileInputFormatTableSource, TestFilterableTableSource, TestInputFormatTableSource, TestNestedProjectableTableSource, TestPartitionableSourceFactory, TestProjectableTableSource, TestTableSourceSinks}
-import org.apache.flink.table.runtime.types.TypeInfoDataTypeConverter
+import org.apache.flink.table.planner.utils._
 import org.apache.flink.types.Row
+
 import org.junit.{Before, Test}
-import java.io.FileWriter
+
 import java.lang.{Boolean => JBool, Integer => JInt, Long => JLong}
-import java.math.{BigDecimal => JDecimal}
-import java.time.{Instant, LocalDateTime, ZoneId}
-
-import org.apache.flink.table.api.internal.TableEnvironmentInternal
-
-import scala.collection.mutable
 
 class TableSourceITCase extends BatchTestBase {
 
@@ -43,18 +34,36 @@ class TableSourceITCase extends BatchTestBase {
   override def before(): Unit = {
     super.before()
     env.setParallelism(1) // set sink parallelism to 1
-    val tableSchema = TableSchema.builder().fields(
-      Array("a", "b", "c"),
-      Array(DataTypes.INT(), DataTypes.BIGINT(), DataTypes.STRING())).build()
-    tEnv.asInstanceOf[TableEnvironmentInternal].registerTableSourceInternal(
-      "MyTable", new TestProjectableTableSource(
-      true,
-      tableSchema,
-      new RowTypeInfo(
-        tableSchema.getFieldDataTypes.map(TypeInfoDataTypeConverter.fromDataTypeToTypeInfo),
-        tableSchema.getFieldNames),
-      TestData.smallData3)
-    )
+    val myTableDataId = TestValuesTableFactory.registerData(TestData.smallData3)
+    tEnv.executeSql(
+      s"""
+        |CREATE TABLE MyTable (
+        |  `a` INT,
+        |  `b` BIGINT,
+        |  `c` STRING
+        |) WITH (
+        |  'connector' = 'values',
+        |  'data-id' = '$myTableDataId',
+        |  'bounded' = 'true'
+        |)
+        |""".stripMargin)
+
+    val filterableTableDataId = TestValuesTableFactory.registerData(
+      TestFilterableTableSource.defaultRows)
+    // TODO: [FLINK-17425] support filter pushdown for TestValuesTableSource
+    tEnv.executeSql(
+      s"""
+         |CREATE TABLE FilterableTable (
+         |  name STRING,
+         |  id BIGINT,
+         |  amount INT,
+         |  price DOUBLE
+         |) WITH (
+         |  'connector' = 'values',
+         |  'data-id' = '$filterableTableDataId',
+         |  'bounded' = 'true'
+         |)
+         |""".stripMargin)
   }
 
   @Test
@@ -101,32 +110,27 @@ class TableSourceITCase extends BatchTestBase {
         Row.of("Betty", new JInt(30000)),
         "Liz"))
 
-    val nested1 = new RowTypeInfo(
-      Array(Types.STRING, Types.INT).asInstanceOf[Array[TypeInformation[_]]],
-      Array("name", "value")
-    )
+    val dataId = TestValuesTableFactory.registerData(data)
 
-    val nested2 = new RowTypeInfo(
-      Array(Types.INT, Types.BOOLEAN).asInstanceOf[Array[TypeInformation[_]]],
-      Array("num", "flag")
-    )
-
-    val deepNested = new RowTypeInfo(
-      Array(nested1, nested2).asInstanceOf[Array[TypeInformation[_]]],
-      Array("nested1", "nested2")
-    )
-
-    val tableSchema = new TableSchema(
-      Array("id", "deepNested", "nested", "name"),
-      Array(Types.LONG, deepNested, nested1, Types.STRING))
-
-    val returnType = new RowTypeInfo(
-      Array(Types.LONG, deepNested, nested1, Types.STRING).asInstanceOf[Array[TypeInformation[_]]],
-      Array("id", "deepNested", "nested", "name"))
-
-    tEnv.asInstanceOf[TableEnvironmentInternal].registerTableSourceInternal(
-      "T",
-      new TestNestedProjectableTableSource(true, tableSchema, returnType, data))
+    // TODO: [FLINK-17428] support nested project for TestValuesTableSource
+    val ddl =
+      s"""
+         |CREATE TABLE T (
+         |  id BIGINT,
+         |  deepNested ROW<
+         |     nested1 ROW<name STRING, `value` INT>,
+         |     nested2 ROW<num INT, flag BOOLEAN>
+         |   >,
+         |   nested ROW<name STRING, `value` INT>,
+         |   name STRING,
+         |   lower_name AS LOWER(name)
+         |) WITH (
+         |  'connector' = 'values',
+         |  'data-id' = '$dataId',
+         |  'bounded' = 'true'
+         |)
+         |""".stripMargin
+    tEnv.executeSql(ddl)
 
     checkResult(
       """
@@ -134,23 +138,19 @@ class TableSourceITCase extends BatchTestBase {
         |    deepNested.nested1.name AS nestedName,
         |    nested.`value` AS nestedValue,
         |    deepNested.nested2.flag AS nestedFlag,
-        |    deepNested.nested2.num AS nestedNum
+        |    deepNested.nested2.num AS nestedNum,
+        |    lower_name
         |FROM T
       """.stripMargin,
-      Seq(row(1, "Sarah", 10000, true, 1000),
-        row(2, "Rob", 20000, false, 2000),
-        row(3, "Mike", 30000, true, 3000)
+      Seq(row(1, "Sarah", 10000, true, 1000, "mary"),
+        row(2, "Rob", 20000, false, 2000, "bob"),
+        row(3, "Mike", 30000, true, 3000, "liz")
       )
     )
   }
 
   @Test
   def testTableSourceWithFilterable(): Unit = {
-    TestFilterableTableSource.createTemporaryTable(
-      tEnv,
-      TestFilterableTableSource.defaultSchema,
-      "FilterableTable",
-      isBounded = true)
     checkResult(
       "SELECT id, name FROM FilterableTable WHERE amount > 4 AND price < 9",
       Seq(
@@ -163,12 +163,6 @@ class TableSourceITCase extends BatchTestBase {
 
   @Test
   def testTableSourceWithFunctionFilterable(): Unit = {
-    TestFilterableTableSource.createTemporaryTable(
-      tEnv,
-      TestFilterableTableSource.defaultSchema,
-      "FilterableTable",
-      isBounded = true,
-      filterableFields = List("amount", "name"))
     checkResult(
       "SELECT id, name FROM FilterableTable " +
         "WHERE amount > 4 AND price < 9 AND upper(name) = 'RECORD_5'",
@@ -178,64 +172,23 @@ class TableSourceITCase extends BatchTestBase {
   }
 
   @Test
-  def testTableSourceWithPartitionable(): Unit = {
-    TestPartitionableSourceFactory.createTemporaryTable(tEnv, "PartitionableTable", true)
-    checkResult(
-      "SELECT * FROM PartitionableTable WHERE part2 > 1 and id > 2 AND part1 = 'A'",
-      Seq(row(3, "John", "A", 2), row(4, "nosharp", "A", 2))
-    )
-  }
-
-  @Test
-  def testCsvTableSource(): Unit = {
-    TestTableSourceSinks.createPersonCsvTemporaryTable(tEnv, "csvTable")
-    checkResult(
-      "SELECT id, `first`, `last`, score FROM csvTable",
-      Seq(
-        row(1, "Mike", "Smith", 12.3),
-        row(2, "Bob", "Taylor", 45.6),
-        row(3, "Sam", "Miller", 7.89),
-        row(4, "Peter", "Smith", 0.12),
-        row(5, "Liz", "Williams", 34.5),
-        row(6, "Sally", "Miller", 6.78),
-        row(7, "Alice", "Smith", 90.1),
-        row(8, "Kelly", "Williams", 2.34)
-      )
-    )
-  }
-
-  @Test
-  def testLookupJoinCsvTemporalTable(): Unit = {
-    TestTableSourceSinks.createOrdersCsvTemporaryTable(tEnv, "orders")
-    TestTableSourceSinks.createRatesCsvTemporaryTable(tEnv, "rates")
-
-    val sql =
-      """
-        |SELECT o.amount, o.currency, r.rate
-        |FROM (SELECT *, PROCTIME() as proc FROM orders) AS o
-        |JOIN rates FOR SYSTEM_TIME AS OF o.proc AS r
-        |ON o.currency = r.currency
-      """.stripMargin
-
-    checkResult(
-      sql,
-      Seq(
-        row(2, "Euro", 119),
-        row(1, "US Dollar", 102),
-        row(50, "Yen", 1),
-        row(3, "Euro", 119),
-        row(5, "US Dollar", 102)
-      )
-    )
-  }
-
-  @Test
   def testInputFormatSource(): Unit = {
-    val tableSchema = TableSchema.builder().fields(
-      Array("a", "b", "c"),
-      Array(DataTypes.INT(), DataTypes.BIGINT(), DataTypes.STRING())).build()
-    TestInputFormatTableSource.createTemporaryTable(
-      tEnv, tableSchema, TestData.smallData3, "MyInputFormatTable")
+    val dataId = TestValuesTableFactory.registerData(TestData.smallData3)
+    tEnv.executeSql(
+      s"""
+         |CREATE TABLE MyInputFormatTable (
+         |  `a` INT,
+         |  `b` BIGINT,
+         |  `c` STRING
+         |) WITH (
+         |  'connector' = 'values',
+         |  'data-id' = '$dataId',
+         |  'bounded' = 'true',
+         |  'runtime-source' = 'InputFormat'
+         |)
+         |""".stripMargin
+    )
+
     checkResult(
       "SELECT a, c FROM MyInputFormatTable",
       Seq(
@@ -246,88 +199,58 @@ class TableSourceITCase extends BatchTestBase {
   }
 
   @Test
-  def testMultiTypeSource(): Unit = {
-    val tableSchema = TableSchema.builder().fields(
-      Array("a", "b", "c", "d", "e", "f"),
-      Array(
-        DataTypes.INT(),
-        DataTypes.DECIMAL(5, 2),
-        DataTypes.VARCHAR(5),
-        DataTypes.CHAR(5),
-        DataTypes.TIMESTAMP(9),
-        DataTypes.TIMESTAMP_WITH_LOCAL_TIME_ZONE(9)
-      )
-    ).build()
-
-    val ints = List(1, 2, 3, 4, null)
-    val decimals = List(
-      new JDecimal(5.1), new JDecimal(6.1), new JDecimal(7.1), new JDecimal(8.123), null)
-    val varchars = List("1", "12", "123", "1234", null)
-    val chars = List("1", "12", "123", "1234", null)
-    val datetimes = List(
-      LocalDateTime.of(1969, 1, 1, 0, 0, 0, 123456789),
-      LocalDateTime.of(1970, 1, 1, 0, 0, 0, 123456000),
-      LocalDateTime.of(1971, 1, 1, 0, 0, 0, 123000000),
-      LocalDateTime.of(1972, 1, 1, 0, 0, 0, 0),
-      null)
-
-    val instants = new mutable.MutableList[Instant]
-    for (i <- datetimes.indices) {
-      if (datetimes(i) == null) {
-        instants += null
-      } else {
-        // Assume the time zone of source side is UTC
-        instants += datetimes(i).toInstant(ZoneId.of("UTC").getRules.getOffset(datetimes(i)))
-      }
-    }
-
-    val data = new mutable.MutableList[Row]
-
-    for (i <- ints.indices) {
-      data += row(
-        ints(i), decimals(i), varchars(i), chars(i), datetimes(i), instants(i))
-    }
-
-    TestDataTypeTableSource.createTemporaryTable(tEnv, tableSchema, "MyInputFormatTable", data.seq)
-
-    checkResult(
-      "SELECT a, b, c, d, e, f FROM MyInputFormatTable",
-      Seq(
-        row(1, "5.10", "1", "1",
-          "1969-01-01T00:00:00.123456789",
-          "1969-01-01T00:00:00.123456789Z"),
-        row(2, "6.10", "12", "12",
-          "1970-01-01T00:00:00.123456",
-          "1970-01-01T00:00:00.123456Z"),
-        row(3, "7.10", "123", "123",
-          "1971-01-01T00:00:00.123",
-          "1971-01-01T00:00:00.123Z"),
-        row(4, "8.12", "1234", "1234",
-          "1972-01-01T00:00",
-          "1972-01-01T00:00:00Z"),
-        row(null, null, null, null, null, null))
+  def testAllDataTypes(): Unit = {
+    val dataId = TestValuesTableFactory.registerData(TestData.fullDataTypesData)
+    tEnv.executeSql(
+      s"""
+         |CREATE TABLE T (
+         |  `a` BOOLEAN,
+         |  `b` TINYINT,
+         |  `c` SMALLINT,
+         |  `d` INT,
+         |  `e` BIGINT,
+         |  `f` FLOAT,
+         |  `g` DOUBLE,
+         |  `h` DECIMAL(5, 2),
+         |  `i` VARCHAR(5),
+         |  `j` CHAR(5),
+         |  `k` DATE,
+         |  `l` TIME(0),
+         |  `m` TIMESTAMP(9),
+         |  `n` TIMESTAMP(9) WITH LOCAL TIME ZONE,
+         |  `o` ARRAY<BIGINT>,
+         |  `p` ROW<f1 BIGINT, f2 STRING, f3 DOUBLE>,
+         |  `q` MAP<STRING, INT>
+         |) WITH (
+         |  'connector' = 'values',
+         |  'data-id' = '$dataId',
+         |  'bounded' = 'true'
+         |)
+         |""".stripMargin
     )
-  }
 
-  @Test
-  def testMultiPaths(): Unit = {
-    val tmpFile1 = TEMPORARY_FOLDER.newFile("tmpFile1.tmp")
-    new FileWriter(tmpFile1).append("t1\n").append("t2\n").close()
-
-    val tmpFile2 = TEMPORARY_FOLDER.newFile("tmpFile2.tmp")
-    new FileWriter(tmpFile2).append("t3\n").append("t4\n").close()
-
-    val schema = new TableSchema(Array("a"), Array(Types.STRING))
-    val paths = Array(tmpFile1.getPath, tmpFile2.getPath)
-
-    TestFileInputFormatTableSource.createTemporaryTable(tEnv, schema, "MyMultiPathTable", paths)
     checkResult(
-      "select * from MyMultiPathTable",
+      "SELECT * FROM T",
       Seq(
-        row("t1"),
-        row("t2"),
-        row("t3"),
-        row("t4")
+        row(
+          true, 127, 32767, 2147483647, 9223372036854775807L, "-1.123", "-1.123", "5.10",
+          1, 1, "1969-01-01", "00:00:00.123", "1969-01-01T00:00:00.123456789",
+          "1969-01-01T00:00:00.123456789Z", "[1, 2, 3]", row(1, "a", "2.3"), "{k1=1}"),
+        row(
+          false, -128, -32768, -2147483648, -9223372036854775808L, "3.4", "3.4", "6.10",
+          12, 12, "1970-09-30", "01:01:01.123", "1970-09-30T01:01:01.123456",
+          "1970-09-30T01:01:01.123456Z", "[4, 5]", row(null, "b", "4.56"), "{k4=4, k2=2}"),
+        row(
+          true, 0, 0, 0, 0, "0.12", "0.12", "7.10",
+          123, 123, "1990-12-24", "08:10:24.123", "1990-12-24T08:10:24.123",
+          "1990-12-24T08:10:24.123Z", "[6, null, 7]", row(3, null, "7.86"), "{k3=null}"),
+        row(
+          false, 5, 4, 123, 1234, "1.2345", "1.2345", "8.12",
+          1234, 1234, "2020-05-01", "23:23:23", "2020-05-01T23:23:23",
+          "2020-05-01T23:23:23Z", "[8]", row(4, "c", null), "{null=3}"),
+        row(
+          null, null, null, null, null, null, null, null, null, null, null, null, null,
+          null, null, null, null)
       )
     )
   }
